+++ src/onegov.election_day/onegov/election_day/forms/notification.py
@@ -11,7 +11,8 @@ class TriggerNotificationForm(Form):
         choices=[],
         validators=[
             InputRequired()
-        ]
+        ],
+        default=['email', 'sms', 'webhooks']
     )
 
     def on_request(self):
@@ -20,13 +21,9 @@ class TriggerNotificationForm(Form):
         principal = self.request.app.principal
 
         self.notifications.choices = []
-        self.notifications.data = []
         if principal.email_notification:
             self.notifications.choices.append(('email', _("Email")))
-            self.notifications.data.append('email')
         if principal.sms_notification:
             self.notifications.choices.append(('sms', _("SMS")))
-            self.notifications.data.append('sms')
         if principal.webhooks:
             self.notifications.choices.append(('webhooks', _("Webhooks")))
-            self.notifications.data.append('webhooks')